updates example images + moves validation to it's own helper file + adds validation for larger date ranges
