+++ views/index.hbs
@@ -1,4 +1,4 @@
-<h1>FT News Barcode</h1>
+<h1>FT News Barcode [Beta]</h1>
 
 <h2>API URL generator</h2>
 
@@ -64,7 +64,7 @@
 </ul>
 
   <br />
-  e.g. <a href="/barcode?width=1024&height=768&dateFrom=2018-11-13&dateTo=2018-11-14&orientation=v&fit=cover">/barcode?width=1024&height=768&dateFrom=2018-11-13&dateTo=2018-11-14&orientation=v&fit=cover</a>
+  e.g. <a href="/barcode?width=1024&height=768&dateFrom=2018-11-13&dateTo=2018-11-14&orientation=v&fit=fill">/barcode?width=1024&height=768&dateFrom=2018-11-13&dateTo=2018-11-14&orientation=v&fit=fill</a>
 </p>
 
 <script>
